requirements.txt
python-jobspy==1.1.69
sendgrid==6.11.0
pandas==2.2.3
requests==2.31.0
linkedin-api==2.0.0
python-dotenv==1.0.0